Display file name when XML load fails
Merge-AllSysmonXml will display an error if the XML file load fails. If the error is due to malformed XML, the error message will include the line number and details. However, the error message does not include the file name. This can make troubleshooting difficult when using sysmon-modular with custom modules.

This change adds a second error message that includes the file name when the XML file load fails.